+++ vars/jenkinsMain.groovy
@@ -20,12 +20,7 @@ def call(){
                 agent { label "builder.ci.jenkins"}
                 steps {
                     script {
-                        cleanWs()
-                        unstash "workspace"
-                        def resourceContent = libraryResource("scripts/post-checkout.sh")
-                        writeFile(file: "post-checkout.sh", text: resourceContent)
-                        sh "bash post-checkout.sh"
-                        stash "workspace"
+                        postCheckoutStage(pipelineManager)
                     }
                 }
             }
@@ -36,19 +31,7 @@ def call(){
                 agent { label "builder.ci.jenkins"}
                 steps {
                     script {
-                        cleanWs()
-                        unstash "workspace"
-                        error "Unstable, exiting now..."
-                        withEnv(["GOPATH=$WORKSPACE", "GOBIN=$GOPATH/bin"]) {
-                            sh "mkdir src bin && go get ./..."
-                            env.PATH="${env.GOPATH}/bin:$PATH"
-                            String repoName = scm.getUserRemoteConfigs()[0].getUrl().tokenize('/').last().split("\\.")[0]
-                            sh "go get -d ./pkg/..."
-                            sh "go install"
-                            sh 'go build -o subway main.go'
-                        
-                            stash "workspace"
-                        }
+                        buildStage(pipelineManager)
                     }
                 }
             }
